src/api/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "open-hallucination-index"
version = "0.1.0"
description = "Middleware API for LLM hallucination detection via claim verification against trusted knowledge graphs and vector stores"
readme = "README.md"
license = "MIT"
requires-python = ">=3.12"
authors = [
    { name = "shiftbloom studio.", email = "fabian@shiftbloom.studio" }
]
keywords = ["llm", "hallucination", "verification", "knowledge-graph", "vector-search"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Framework :: FastAPI",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.14",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "fastapi>=0.128.0",
    "uvicorn[standard]>=0.40.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "httpx>=0.28.1",
    "openai>=2.14.0",
    "neo4j>=6.0.3",
    "qdrant-client>=1.16.2",
    "redis>=7.1.0",
    "pyyaml>=6.0.3",
    "mcp>=1.25.0",
    "sentence-transformers>=5.2.0",
    "wikitextparser>=0.56.4",
    "uvloop>=0.22.1; sys_platform != 'win32'",
    "orjson>=3.11.5"
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.10",
    "mypy>=1.19.1",
    "pre-commit>=4.5.1",
]

[project.scripts]
ohi-server = "config.entrypoint:main"

[tool.hatch.build.targets.wheel]
packages = ["models", "interfaces", "pipeline", "adapters", "server", "config", "services"]

[tool.ruff]
target-version = "py314"
line-length = 100
src = [".", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "ASYNC"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.14"
strict = true
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = """
    --cov=models
    --cov=interfaces
    --cov=pipeline
    --cov=adapters
    --cov=server
    --cov=config
    --cov=services
    --cov-report=xml
    --cov-report=term-missing
    --cov-fail-under=70
    -v
"""
